.claude/skills/fastapi-backend/backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Configuration and environment
pydantic-settings==2.0.3
python-dotenv==1.0.0

# Data validation and serialization
pydantic==2.5.0
pydantic[email]==2.5.0

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Vector database and RAG (placeholders - replace with actual implementations)
# qdrant-client==1.7.0
# langchain==0.0.350
# langchain-openai==0.0.2

# Database (placeholder - replace with actual database)
# sqlalchemy==2.0.23
# asyncpg==0.29.0
# alembic==1.13.1

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Development tools
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Optional utilities
python-dateutil==2.8.2
pytz==2023.3